campo-api/requirements.txt
# FastAPI and Web
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
httpx==0.26.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
aiopg==1.4.0
asyncpg==0.29.0

# Auth
python-jose[cryptography]==3.3.0

# Storage
boto3==1.34.10

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.0
python-dateutil==2.8.2

# Image Processing (EXIF, watermark, thumbnails)
Pillow==10.1.0
piexif==1.1.3

# PDF Generation (ReportLab for EVD01)
reportlab==4.0.7
qrcode[pil]==7.4.2

# Observability
prometheus-client==0.19.0
python-json-logger==2.0.7
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Background Jobs
celery==5.3.4
redis==5.0.1
flower==2.0.1  # Celery monitoring
